requirements.txt
bs4
requests
pandas
pyarrow
seaborn
python-bidi
matplotlib
google-cloud-storage